A integração entre a camada data/local (Room) e a camada repository está funcionando corretamente:

+++ app/src/main/java/com/devspace/taskbeats/data/model/SubTaskUiData.kt
@@ -1,8 +1,8 @@
 package com.devspace.taskbeats.data.model
 
 data class SubTaskUiData(
-    val id: Int = 0,
-    val taskId: Int,
+    val id: Long = 0,
+    val taskId: Long,
     val title: String
