+++ app/src/main/java/com/devspace/taskbeats/repository/TaskRepository.kt
@@ -4,8 +4,8 @@ import com.devspace.taskbeats.data.model.SubTaskUiData
 import com.devspace.taskbeats.data.model.TaskUiData
 
 interface TaskRepository {
-    suspend fun getTaskById(taskId: Int): TaskUiData?
-    suspend fun getSubTasksByTaskId(taskId: Int): List<SubTaskUiData>
+    suspend fun getTaskById(taskId: Long): TaskUiData?
+    suspend fun getSubTasksByTaskId(taskId: Long): List<SubTaskUiData>
     suspend fun generateSubTasks(taskTitle: String): List<SubTaskUiData>
-    suspend fun deleteTask(taskId: Int)
+    suspend fun deleteTask(taskId: Long)
 }
